Moved SquareDriver to testing package
